Implement basic math operations in `MathServiceImpl` and update method signatures in `MathService`.

+++ Task1 Simple Calculator/src/MathService.java
@@ -1,7 +1,7 @@
 
 public interface MathService {
-    int add(double firstNumber, double secondNumber);
-    int subtract(double firstNumber, double secondNumber);
-    int multiply(double firstNumber, double secondNumber);
-    int divide(double firstNumber, double secondNumber);
+    double add(double firstNumber, double secondNumber);
+    double subtract(double firstNumber, double secondNumber);
+    double multiply(double firstNumber, double secondNumber);
+    double divide(double firstNumber, double secondNumber);
 }
